requirements.txt
asgiref==3.8.0
certifi==2024.2.2
distlib==0.3.8
Django==5.0.3
django-filter==24.1
djangorestframework==3.15.0
filelock==3.14.0
Markdown==3.6
pipenv==2023.12.1
platformdirs==4.2.2
setuptools==70.0.0
sqlparse==0.4.4
virtualenv==20.26.2
